Add support for load-balancer.hetzner.cloud/external-dns-hostname annotation
When this annotation is added to a Service, the controller will apply two annotations for external-dns:

1) external-dns.alpha.kubernetes.io/target: <ipv4-address>,<ipv6-address>
2) external-dns.alpha.kubernetes.io/hostname: the value from load-balancer.hetzner.cloud/external-dns-hostname

The value from load-balancer.hetzner.cloud/external-dns-hostname will be added to svc.status instead of IPv4 and IPv6 addresses.

This is useful and convenient for proxy-protocol mode. By default, Cilium, kube-proxy (IPVS), and kube-router (IPVS) forward traffic for the LoadBalancer IP to backend pods. However, if svc.status contains a hostname instead of IP addresses, they do not perform this forwarding.

Additionally, external-dns will automatically create the corresponding DNS record.